Issue #385 Move config base and MCVirt config into subdir
