+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/auth/auth.py
@@ -21,7 +21,7 @@ import os
 import Pyro4
 from texttable import Texttable
 
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.exceptions import (UserNotPresentInGroup, InsufficientPermissionsException,
                                UnprivilegedUserException, InvalidPermissionGroupException,
                                DuplicatePermissionException,
