+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/auth/factory.py
@@ -19,7 +19,7 @@
 
 from texttable import Texttable
 
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.exceptions import (IncorrectCredentials, InvalidUsernameException,
                                UserDoesNotExistException, InvalidUserTypeException,
                                UserAlreadyExistsException, BlankPasswordException)
