+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/auth/group/factory.py
@@ -24,7 +24,7 @@ from mcvirt.rpc.expose_method import Expose, Transaction
 from mcvirt.auth.permissions import PERMISSIONS
 from mcvirt.auth.group.group import Group
 from mcvirt.argument_validator import ArgumentValidator
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.exceptions import (GroupAlreadyExistsError,
                                GroupDoesNotExistError)
