+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/auth/user_types/cluster_user.py
@@ -18,7 +18,7 @@
 # along with MCVirt.  If not, see <http://www.gnu.org/licenses/>
 
 from mcvirt.auth.user_types.user_base import UserBase
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.rpc.expose_method import Expose
