+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/node/node.py
@@ -19,7 +19,7 @@
 
 import Pyro4
 
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.auth.permissions import PERMISSIONS
 from mcvirt.rpc.pyro_object import PyroObject
 from mcvirt.rpc.expose_method import Expose
