+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/test/node/node_tests.py
@@ -18,7 +18,7 @@
 import unittest
 
 from mcvirt.exceptions import MCVirtTypeError
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.test.test_base import TestBase
