+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/thread/watchdog.py
@@ -22,7 +22,7 @@ from mcvirt.thread.repeat_timer import RepeatTimer
 from mcvirt.rpc.pyro_object import PyroObject
 from mcvirt.syslogger import Syslogger
 from mcvirt.argument_validator import ArgumentValidator
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.rpc.expose_method import Expose
 from mcvirt.utils import dict_merge
 from mcvirt.auth.permissions import PERMISSIONS
