+++ source/mcvirt-daemon/usr/lib/python2.7/dist-packages/mcvirt/virtual_machine/virtual_machine.py
@@ -41,7 +41,7 @@ from mcvirt.exceptions import (MigrationFailureExcpetion, InsufficientPermission
                                DeleteProtectionEnabledError)
 from mcvirt.syslogger import Syslogger
 from mcvirt.virtual_machine.agent_connection import AgentConnection
-from mcvirt.mcvirt_config import MCVirtConfig
+from mcvirt.config.mcvirt_config import MCVirtConfig
 from mcvirt.virtual_machine.disk_drive import DiskDrive
 from mcvirt.virtual_machine.usb_device import UsbDevice
 from mcvirt.virtual_machine.virtual_machine_config import VirtualMachineConfig
